+++ web/admin/src/views/dashboard/feed_viewer/feed_view_container.vue
